Separate out testing that getting a job before creating does not work nicely
